server/requirements.txt
anyio==4.4.0
boto3==1.35.5
botocore==1.35.5
certifi==2024.7.4
charset-normalizer==3.3.2
Cython==3.0.11
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.8
jmespath==1.0.1
openrank-sdk==0.2.2
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
urllib3==2.2.2
python-dotenv==0.19.1
